perf: cut template tree and Swiper from public bundle
Trim ~915 KB of unused JavaScript from public pages by isolating
builder-only dependencies behind dedicated modules and skipping
runtime initializers that have nothing to do.

- Stop calling useCanvasSlider from LayerRendererPublic (it never
  ran in public mode but was still pulling Swiper into every page)
- Extract slider/lightbox constants and isSliderLayerName into
  lib/slider-constants.ts so consumers no longer transitively load
  the 1.5 MB lib/templates/* tree
- Move getLayerIcon/getLayerName into lib/layer-display-utils.ts
  so lib/layer-utils.ts (used by the public renderer) is template
  free; builder callers are repointed to the new module
- Move pure runtime translation helpers into lib/locale-runtime.ts
  and have lib/page-utils.ts import from there to break the
  localisation -> templates chain
- Only mount AnimationInitializer when a page actually contains
  layer interactions, keeping GSAP off static pages
- Plumb lcpCandidateLayerId through LayerRendererPublic so the
  earlier LCP prioritization commit also applies to the public
  renderer
- Replace global lodash import with lodash/cloneDeep in
  layer-utils for better tree shaking
