+++ app/(builder)/ycode/components/RichTextLinkSettings.tsx
@@ -35,7 +35,8 @@ import { useAssetsStore } from '@/stores/useAssetsStore';
 import { useEditorStore } from '@/stores/useEditorStore';
 import { getAssetIcon } from '@/lib/asset-utils';
 import { collectionsApi } from '@/lib/api';
-import { getLayerIcon, getLayerName, getCollectionVariable, findLayersWithAnchorId } from '@/lib/layer-utils';
+import { getCollectionVariable, findLayersWithAnchorId } from '@/lib/layer-utils';
+import { getLayerIcon, getLayerName } from '@/lib/layer-display-utils';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
 import PageSelector from './PageSelector';
 import { filterFieldGroupsByType, flattenFieldGroups, LINK_FIELD_TYPES, buildReferenceItemOptions, type FieldGroup } from '@/lib/collection-field-utils';
