+++ lib/slider-constants.ts
@@ -1,0 +1,79 @@
+/**
+ * Slider / Lightbox constants and pure type guards.
+ *
+ * Lives in its own module (with zero dependencies on `lib/templates/*`) so
+ * that the public renderer and other lightweight consumers can import these
+ * symbols without dragging the entire ~1.5 MB template tree
+ * (`lib/templates/layouts.ts` + friends) into their bundle.
+ */
+
+import type { SliderSettings, LightboxSettings } from '@/types';
+
+/** Default slider settings applied when creating a new slider */
+export const DEFAULT_SLIDER_SETTINGS: SliderSettings = {
+  navigation: true,
+  groupSlide: 1,
+  slidesPerGroup: 1,
+  loop: 'none',
+  centered: false,
+  touchEvents: false,
+  slideToClicked: false,
+  mousewheel: false,
+
+  pagination: true,
+  paginationType: 'bullets',
+  paginationClickable: true,
+  autoplay: false,
+  pauseOnHover: true,
+  delay: '3',
+  animationEffect: 'slide',
+  easing: 'ease-in-out',
+  duration: '0.5',
+};
+
+/** Default lightbox settings applied when creating a new lightbox */
+export const DEFAULT_LIGHTBOX_SETTINGS: LightboxSettings = {
+  files: [],
+  filesSource: 'files',
+  filesField: null,
+  thumbnails: true,
+  navigation: true,
+  pagination: true,
+  zoom: false,
+  doubleTapZoom: false,
+  mousewheel: false,
+  overlay: 'light',
+  groupId: '',
+  animationEffect: 'slide',
+  easing: 'ease-in-out',
+  duration: '0.5',
+};
+
+/** All layer names that are part of the slider element */
+export const SLIDER_LAYER_NAMES = [
+  'slider', 'slides', 'slide',
+  'slideNavigationWrapper', 'slideButtonPrev', 'slideButtonNext',
+  'slidePaginationWrapper', 'slideBullets', 'slideFraction', 'slideBullet',
+] as const;
+
+export type SliderLayerName = typeof SLIDER_LAYER_NAMES[number];
+
+/** Check if a layer name belongs to the slider element family */
+export function isSliderLayerName(name: string): name is SliderLayerName {
+  return (SLIDER_LAYER_NAMES as readonly string[]).includes(name);
+}
+
+/** Swiper CSS classes needed for core layout (used on canvas + production) */
+export const SWIPER_CLASS_MAP: Record<string, string> = {
+  slider: 'swiper',
+  slides: 'swiper-wrapper',
+  slide: 'swiper-slide',
+};
+
+/** Data attributes added to slider nav/pagination elements on production for Swiper targeting */
+export const SWIPER_DATA_ATTR_MAP: Record<string, string> = {
+  slideButtonPrev: 'data-slider-prev',
+  slideButtonNext: 'data-slider-next',
+  slideBullets: 'data-slider-pagination',
+  slideFraction: 'data-slider-fraction',
+};
